When the HTS221 device isn't available don't crash the server
In some cases where the HTS221 sensor is optional we don't what to crash
a firmware's application start due to it missing.

This update will by default gracefully stop the `HTS221.Server` so that
firmwares that use this library will continue to work if the sensor is
not on the device. This does log that we tried to use the sensor but we
were unable to find it.